reafctoring : remove sercah grid to seperate file
